requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.4.0
requests==2.28.2
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
